requirements.txt
fastapi
uvicorn[standard]
openai
requests
PyPDF2
python-docx
pytesseract
Pillow
